Repair
Fix the problem of the read.csv.

Now we have a function DS that control the robot in position to follow a simple DS.
And the function Send_post that control the robot in position to foolows a trjaectory from a .csv

But the implementation of the package trak_ik is not working actually.